Create the methods for the demage of an enemy and not pass the map from controller but with the map that has been created in the starter.

+++ src/sgf/managers/EnemyManager.java
@@ -17,7 +17,7 @@ public interface EnemyManager {
      * Inflicts damage on the enemy.
      * @param damage the amount of inflicted damage
      */
-    void damage(int damage);
+    void damage(double damage);
 
     /**
      * Stops the thread.
